+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes
 - Removed usage of NodeJS modules from the SDK and some dependencies. With this change, the SDK should now work with some of the latest frameworks that use the latest versions of bundlers such as Vite and Webpack.
 - Removed unnecessary files from the generated npm package.
 - Links to source maps are now included in the generated npm package.
+- Migrated the `AudioPlayer` dependency and is now part of the SDK. This change fixes an issue where source maps are not properly loaded.
 
 2.5.0 (May 9, 2023)
 ===================